swapped bigint dates to number (integer) as int enough for milliseconds

+++ apps/api/src/modules/auth/utils/jwt.util.ts
@@ -61,7 +61,7 @@ export function createJwtUtil(config: Config) {
   }
 
   function getRefreshTokenExpiresAt() {
-    return BigInt(Date.now() + config.JWT_REFRESH_EXPIRY_SECONDS * 1000)
+    return Date.now() + config.JWT_REFRESH_EXPIRY_SECONDS * 1000
   }
 
   return {
